handling of textareas with html5 required-attribute
tinyMCE doesn't work with textareas with a required-attribute, since the
constraint-validation that is triggered by the required-attribute will
prevent the 'submit'-event on the form which normally triggers tinyMCE
to save the editor-contents into the hidden textarea.

To circumvent this problem, an onchange-handler is attached to textareas with
the required-attribute that will update the textarea contents from the
editor when the focus leaves the editor (which is at least once before
the constraint-validation starts.

(fixes #18)